Refactor budget expense group drag-and-drop handling
Extracted drag-and-drop logic for reordering budget expense groups into a dedicated BudgetExpenseGroupReorderHandler class. Updated BudgetViewModel to use this handler and exposed it for binding in the view. Modified BudgetPage.xaml to bind drag-and-drop handlers to the new handler, improving separation of concerns and maintainability.
